refactor collector into own package
