+++ Dockerfile
@@ -11,6 +11,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 ADD cov19 /app/cov19
 COPY web.py /app
+COPY data_tracker.py /app
 COPY gunicorn_start.sh /app
 
 CMD ["./gunicorn_start.sh"]
