+++ cov19/__init__.py
@@ -1,4 +1,5 @@
 from .statistics import Cov19Statistics
-from .collect import Germany, Austria, Switzerland, UnitedKingdom, UnitedStates
+from .collect import DataCollector, Germany, Austria, Switzerland, UnitedKingdom, UnitedStates
 from .helper import get_query_interval
 from .exceptions import RequestError
+
