+++ cov19/collect/__init__.py
@@ -1,0 +1,6 @@
+from .collector import DataCollector
+from .at import Austria
+from .de import Germany
+from .ch import Switzerland
+from .uk import UnitedKingdom
+from .us import UnitedStates
